# feat: Make generate_layers_json.sh 73x faster