<!--- Please remember to review the [contribution guidelines](https://github.com/DataDog/serverless-plugin-datadog/blob/master/CONTRIBUTING.md) if you have not yet done so._  --->

### Problem
The script `generate_layers_json.sh` is slow. On my machine, it takes:
- 16min49s to generate `layers.json`
- 46s to generate `layers-gov.json`

These steps take most of the time of running `publish_prod.sh` for releasing a version of Serverless Plugin. Partially due to this, people are reluctant to publish new versions of Plugin, making it lag behind on extension and tracing layers, resulting in customer complaints such as:
- https://github.com/DataDog/serverless-plugin-datadog/issues/569
- https://github.com/DataDog/serverless-plugin-datadog/issues/564
<!--- What inspired you to submit this pull request? --->

### What does this PR achieve?
Reduces the time to run this script.
- For generating `layers.json`, reduces the time from **16m49s** to **12s** (**-98.8%, 87x**)
- For generating `layers-gov.json`, reduces the time from **41s** to **2.4s** (**-94%, 17x**)
- Total: **17m30s** to **~14.4s** (**-98.6%, 73x**)

### How does this PR achieve it?
Before, the script makes an synchronous API call for each <region, layer_name> pair. The number of pairs was relatively low back in April 2019: [15 regions * 5 layer_names = 75 pairs](https://github.com/DataDog/serverless-plugin-datadog/commit/4e686e5571186b82fc5a6aa8c72bd71a110afc6e). However, today we have 32 regions * 26 layer_names = 832 pairs (~11x). The network delay for the 832 sequential synchronous calls add up to a lot.

In this PR:
1. For each region, we make a single API call, which fetches all the layers in that region, then store the result to a file.
2. Perform the 32 API calls (for 32 regions) in the last step in parallel.

Thus network delay is no longer a bottleneck.
<!--- A brief description of the change being made with this pull request. --->

### Testing Guidelines
#### Steps
1. Generate `layers.json` by running `aws-vault exec sso-prod-engineering -- ./scripts/generate_layers_json.sh`
2. Generate `layers-gov.json` by running `aws-vault exec sso-govcloud-us1-fed-engineering -- ./scripts/generate_layers_json.sh -g`

#### Result
1. The json files are not changed.
2. Scripts are faster.

Demo for generating `layers.json`:
Before:
![generate_layers_before](https://github.com/user-attachments/assets/278d61e5-7ed3-4f60-acc5-d6c69627f4f9)

After:
![generate_layers_after](https://github.com/user-attachments/assets/ab95cd5b-c183-42c5-a683-2830db2a232d)

<!--- How did you test this pull request? --->

### Additional Notes

<!--- Anything else we should know when reviewing? --->

### Types of changes

- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Misc (docs, refactoring, dependency upgrade, etc.)

### Check all that apply

- [x] This PR's description is comprehensive
- [ ] This PR contains breaking changes that are documented in the description
- [ ] This PR introduces new APIs or parameters that are documented and unlikely to change in the foreseeable future
- [ ] This PR impacts documentation, and it has been updated (or a ticket has been logged)
- [ ] This PR's changes are covered by the automated tests
- [ ] This PR collects user input/sensitive content into Datadog